src/requirements.txt
djangorestframework>=3.12.2,<3.2
django-oauth-toolkit>=1.3.3,<1.4
django-cors-middleware>=1.5.0,<1.6